Reuse getClientDetails
